Move setup() method into separate file
There will be so much more to setup, so the better isolated it is, the
better off we will be.
